Cambios menores.

+++ src/Mundo.h
@@ -32,7 +32,7 @@ class Mundo
 
 	int nivel;
 	float t_DisparoDoble; //tiempo restante del disparo doble
-	float puntuacion; //Puntuación del jugador
+	int puntuacion; //Puntuación del jugador
 	bool pressed_spacebar; //Variable auxiliar para gestionar flancos negativos en tecla espacio
 
 public:
